+++ transcriptic/util.py
@@ -4,6 +4,7 @@ from past.builtins import basestring
 from past.utils import old_div
 import re
 import itertools
+import sys
 
 
 def natural_sort(l):
@@ -141,3 +142,19 @@ def humanize(well_ref, well_count, col_count):
 def by_well(datasets, well):
     return [datasets[reading].props['data'][well][0] for
             reading in list(datasets.keys())]
+
+
+def makedirs(name, mode=None, exist_ok=False):
+    """Forward ports `exist_ok` flag for Py2 makedirs. Retains mode defaults"""
+    from os import makedirs
+    if sys.version_info[0] < 3:
+        # Note that Py2 makedirs still errors on all errno.EEXIST, unlike Py3
+        mode = mode if mode is not None else 0777
+        try:
+            makedirs(name, mode)
+        except OSError:
+            if not exist_ok:
+                raise
+    else:
+        mode = mode if mode is not None else 0o777
+        makedirs(name, mode, exist_ok)